src/employee/app/requirements.txt
requests==2.26.0
django>=3.2,<4.0
djangorestframework
django_filter
drf-spectacular
psycopg2-binary
django-autoslug
django-cors-headers
django.dotenv
gunicorn
pika
pynvim
